requirements.txt
alembic==1.18.1
Flask==3.1.2
flask_migrate==4.1.0
flask_sqlalchemy==3.1.1
flask_wtf==1.2.2
injector==0.24.0
langchain_core==1.2.7
langchain_openai==1.1.7
openai==2.15.0
pytest==9.0.2
python-dotenv==1.2.1
WTForms==3.2.1

pydantic